requirements.txt
streamlit>=1.31.0
supabase>=2.9.0
python-dotenv>=1.0.0
pandas>=2.2.0
plotly>=5.18.0
